Updates to the latest version of Goom Adds mmx and ppc assembly optimizations. Adds 5 config entries for this post plugin :   - frames per second   - width   - height   - usage of assembly optimizations   - colorspace conversion method PPc optimizations are not tested. The compiler option -fomit-frame-pointer is disabled.
CVS patchset: 3762
CVS date: 2003/01/03 21:29:20

+++ src/post/goom/ifs.h
@@ -7,28 +7,29 @@
 
 #include "goom_config.h"
 
-typedef struct _ifsPoint {
-  gint16 x,y;
-} IFSPoint;
+typedef struct _ifsPoint
+{
+	gint32  x, y;
+}
+IFSPoint;
 
-/*  init ifs for a (width)x(height) output. */
-void init_ifs (int width, int height);
+/* init ifs for a (width)x(height) output. */
+void    init_ifs (int width, int height);
 
-/*
- *  draw an ifs on the buffer (which size is width * height)
- *  increment means that we draw 1/increment of the ifs's points
- */
-void ifs_update (guint32 *buffer, guint32 *back, int width, int height, int increment);
+/* draw an ifs on the buffer (which size is width * height) */
+/* increment means that we draw 1/increment of the ifs's points */
+void    ifs_update (guint32 * buffer, guint32 * back, int width, int height,
+										int increment);
 
-/*  free all ifs's data. */
-void release_ifs ();
+/* free all ifs's data. */
+void    release_ifs ();
 
 
 /* DONT USE !!! deprecated
  * return a an array of points.
  * WARNING !!! do not free it !!! it also has an internal use..
  */
-IFSPoint * draw_ifs (int * nbPoints);
+IFSPoint *draw_ifs (int *nbPoints);
 
 
 #endif
